refactor(engine): extract model_manager module-level helpers
Move the ~915 lines of module-level helpers out of model_manager.py
(4,028 → ~3,100 lines) into three focused modules:

- model_load_utils.py: config sanitize, tokenizer EOS stops, underscore
  buffer materialization, model-type fallback loader, language-tower
  quantization, Gemma 4 unified-text loader
- cache_capabilities.py: KV-cache serializability/trim/persistence
  predicates
- loaded_model.py: LoadedModel dataclass, structural_copy, load error
  classes, parse_keep_alive

model_manager.py re-exports every moved name, so imports and facade
monkeypatch targets for code still living there are unchanged. Tests
whose patched consumer moved (Gemma 4 dispatch, tokenizer-EOS caplog
loggers) are repointed to the new modules. ModelManager itself stays
put — splitting it into mixins would invalidate the ~135
model_manager.mx/.settings/.gc patch sites wholesale (follow-up).
