+++ olmlx/engine/model_manager.py
@@ -5,17 +5,14 @@ import gc
 import importlib
 import json
 import logging
-import re
 import threading
 import time
-from dataclasses import dataclass, field
 from pathlib import Path
 from typing import TYPE_CHECKING, Any
 
 import mlx.core as mx
-import mlx.nn as nn
 
-from olmlx.config import FlashMoeConfig, SyncMode, experimental as global_experimental
+from olmlx.config import FlashMoeConfig, experimental as global_experimental
 from olmlx.config import resolve_experimental, settings
 from olmlx.engine.registry import (
     _FLASH_MOE_INCOMPATIBLE_STRATEGIES,
@@ -42,6 +39,40 @@ if TYPE_CHECKING:
 
 from olmlx.models.store import _dir_size, _strip_ollama_tag
 
+# Extracted helper modules (#split-large-modules). Re-exported here so
+# `from olmlx.engine.model_manager import X` and facade-level monkeypatch
+# targets keep working for code that still lives in this module.
+from olmlx.engine.model_load_utils import (  # noqa: F401
+    _sanitize_model_config_in_place,
+    _ensure_tokenizer_eos_in_stops,
+    _materialize_module_buffers,
+    _load_with_model_type_fallback,
+    _quantize_language_tower,
+    _load_gemma4_unified_text,
+    _maybe_load_gemma4_unified_text,
+)
+from olmlx.engine.cache_capabilities import (  # noqa: F401
+    _is_serializable_cache,
+    _kv_quant_blocks_snapshot,
+    _TRIMMABLE_CACHE_CLASSES,
+    _cache_supports_trim,
+    _PERSISTABLE_CACHE_CLASSES,
+    _PERSISTABLE_CACHE_CLASSES_WITH_CHECKPOINT,
+    _layer_layout_is_mixed_excluded,
+    _cache_supports_checkpoint_persistence,
+    _cache_supports_persistence,
+)
+from olmlx.engine.loaded_model import (  # noqa: F401
+    ModelLoadTimeoutError,
+    SpectralCalibrationMissingError,
+    ShardCalibrationMissingError,
+    ActiveRequestsError,
+    structural_copy,
+    LoadedModel,
+    _is_cross_encoder_config,
+    parse_keep_alive,
+)
+
 logger = logging.getLogger(__name__)
 
 #: Default max tokens collected per head during SpectralQuant calibration.
@@ -62,921 +93,6 @@ _FALLBACK_EXCEPTIONS = (
     OSError,
     json.JSONDecodeError,
 )
-
-
-def _sanitize_model_config_in_place(load_path) -> None:
-    """Fix known on-disk config.json issues that block transformers loading.
-
-    Currently handles: ``layer_types`` longer than ``num_hidden_layers``.
-    Step-3.5 ships with ``num_hidden_layers=45`` but ``len(layer_types)=48``
-    (the trailing 3 entries describe MTP layers whose weights mlx-lm's
-    sanitize() drops). Newer transformers' ``validate_layer_type`` rejects
-    the mismatch. Truncating ``layer_types`` to match is consistent with
-    what mlx-lm does for the weights and is idempotent on subsequent loads.
-    """
-    config_file = Path(load_path) / "config.json"
-    if not config_file.exists():
-        return
-    try:
-        cfg = json.loads(config_file.read_text())
-    except json.JSONDecodeError:
-        return
-
-    nhl = cfg.get("num_hidden_layers")
-    layer_types = cfg.get("layer_types")
-    if (
-        isinstance(layer_types, list)
-        and isinstance(nhl, int)
-        and nhl > 0
-        and len(layer_types) > nhl
-    ):
-        logger.info(
-            "Truncating layer_types from %d to %d entries in %s "
-            "(num_hidden_layers); excess entries describe layers mlx-lm drops "
-            "(e.g. MTP).",
-            len(layer_types),
-            nhl,
-            config_file,
-        )
-        cfg["layer_types"] = layer_types[:nhl]
-        try:
-            config_file.write_text(json.dumps(cfg, indent=2, ensure_ascii=False))
-        except OSError:
-            logger.warning(
-                "Failed to write sanitized layer_types to %s "
-                "(read-only?); model may fail to load if transformers validates "
-                "the mismatch.",
-                config_file,
-            )
-
-
-def _ensure_tokenizer_eos_in_stops(tokenizer: Any) -> None:
-    """Add the tokenizer's own ``eos_token_id`` to its stop-token set.
-
-    Workaround for repos (e.g. ``mlx-community/Qwen2.5-Coder-1.5B-Instruct-4bit``,
-    issue #308) whose ``config.json`` declares ``eos_token_id`` as a token
-    different from the chat template's real end-of-turn token (the
-    ``eos_token`` field in ``tokenizer_config.json``). mlx-lm's ``load()``
-    feeds only the config.json value into ``TokenizerWrapper.eos_token_ids``,
-    so generation does not stop at the template's actual EOT and the EOT
-    string leaks into the decoded response.
-    """
-    # ``add_eos_token`` is the TokenizerWrapper marker — keep the gate on it
-    # so plain HF tokenizers and mlx-vlm processors are skipped — but mutate
-    # the ``eos_token_ids`` set directly to bypass the wrapper's stringly-typed
-    # API (``add_eos_token(str)`` does ``int(token)`` first then a vocab
-    # lookup; we already have the integer so set-mutation is unambiguous).
-    if not callable(getattr(tokenizer, "add_eos_token", None)):
-        return
-    stops = getattr(tokenizer, "eos_token_ids", None)
-    if not isinstance(stops, set):
-        # Symmetric with the missing-_tokenizer DEBUG log below: an mlx-lm
-        # change from set to list/frozenset/dict for eos_token_ids would
-        # otherwise silently disable the workaround.
-        logger.debug(
-            "TokenizerWrapper.eos_token_ids is %s (not set) on %s; eos "
-            "stop-set augmentation skipped (issue #308).",
-            type(stops).__name__,
-            type(tokenizer).__name__,
-        )
-        return
-    inner_tok = getattr(tokenizer, "_tokenizer", None)
-    if inner_tok is None:
-        # Past the ``add_eos_token`` gate but no ``_tokenizer`` attribute —
-        # mlx-lm likely renamed the field. Log at debug so the regression is
-        # discoverable in DEBUG-enabled environments without spamming the
-        # warning channel for a possibly-deliberate variant.
-        logger.debug(
-            "TokenizerWrapper._tokenizer not accessible on %s; eos stop-set "
-            "augmentation skipped (issue #308).",
-            type(tokenizer).__name__,
-        )
-        return
-    inner_eos = getattr(inner_tok, "eos_token_id", None)
-    if isinstance(inner_eos, list):
-        # Stock HF tokenizers always surface a single int here; defensive
-        # against custom trust_remote_code=True tokenizers that override
-        # ``eos_token_id`` to return list[int].
-        stops.update(t for t in inner_eos if isinstance(t, int))
-        return
-    if inner_eos is None:
-        # Tokenizer has no EOS configured — legitimate for some HF tokenizers
-        # (e.g. base/non-instruction-tuned variants). Silent no-op.
-        return
-    if not isinstance(inner_eos, int):
-        # mlx-lm renamed ``_tokenizer`` or the inner HF tokenizer surfaces an
-        # unexpected type. ``warning``, not ``debug``: this branch indicates
-        # the #308 workaround has silently regressed because mlx-lm changed
-        # its internals, and we recover by no-op'ing — operators need to see
-        # the signal in default logging configs, not only under DEBUG.
-        logger.warning(
-            "Inner eos_token_id has unexpected type %s on %s; skipping "
-            "eos stop-set augmentation (issue #308 workaround may have "
-            "regressed against mlx-lm internals).",
-            type(inner_eos).__name__,
-            type(tokenizer).__name__,
-        )
-        return
-    if inner_eos in stops:
-        return
-    stops.add(inner_eos)
-
-
-def _materialize_module_buffers(model: Any) -> None:
-    """Eager-eval a model's non-parameter array buffers on the current thread.
-
-    mlx-lm's ``load`` materializes ``mx.eval(model.parameters())``, but
-    ``nn.Module.parameters()`` skips **underscore-prefixed** dict keys (the
-    same traversal gap behind the flash-MoE #657 crash). Scaled-RoPE variants
-    — Yarn, longrope, llama3, proportional — precompute ``self._freqs`` (and
-    ``self._scale``) as *lazy* arrays in ``__init__``, so they are never
-    reached by the load-time parameter eval and stay bound to the load
-    thread's Metal stream.
-
-    Under mlx >= 0.31.2 thread-local streams (#499) the model is loaded on one
-    ``asyncio``/generation worker thread and generated on another, so the first
-    forward that evaluates a graph referencing the lazy ``_freqs`` raises
-    "There is no Stream(gpu, N) in current thread" — deterministically, on
-    every request. This surfaces on long-context models whose RoPE is scaled
-    (e.g. the 1M-context Qwen3.5 hybrid ``empero-ai/Qwythos-9B``); models on
-    the default/mrope RoPE carry no ``_freqs`` buffer and are unaffected.
-
-    Materializing the buffers here, on the load thread, turns them into leaves
-    (which carry no stream binding) so they can be read from any worker thread.
-    ``nn.Module`` is a ``dict`` subclass whose attributes are stored as items,
-    so we walk ``named_modules()`` and collect the underscore-keyed array
-    leaves (recursing into plain list/tuple/dict containers, but not into
-    child ``nn.Module``s — those are visited by ``named_modules()`` in turn).
-    """
-    buffers: list[mx.array] = []
-
-    def _collect(value: Any) -> None:
-        if isinstance(value, nn.Module):
-            return  # visited independently by named_modules()
-        if isinstance(value, mx.array):
-            buffers.append(value)
-        elif isinstance(value, (list, tuple)):
-            for item in value:
-                _collect(item)
-        elif isinstance(value, dict):
-            for item in value.values():
-                _collect(item)
-
-    try:
-        modules = model.named_modules()
-    except AttributeError:
-        return  # not an nn.Module tree (e.g. a test double) — nothing to do
-    for _name, module in modules:
-        for key, value in module.items():
-            if key.startswith("_"):
-                _collect(value)
-    if buffers:
-        mx.eval(buffers)
-
-
-def _load_with_model_type_fallback(mlx_lm, load_path, **kwargs):
-    """Load model + tokenizer, remapping unrecognised model_type if needed.
-
-    Some very new models (e.g. DeepSeek V3.2 with model_type "deepseek_v32")
-    aren't in the installed transformers' CONFIG_MAPPING yet, causing
-    ``PreTrainedConfig.__post_init__`` to crash during tokenizer loading.
-
-    The model architecture (loaded by mlx-lm's own registry) is fine — only
-    the tokenizer loading via ``AutoTokenizer`` / ``PreTrainedConfig`` fails.
-    So we load the model first with the real config, then patch config.json
-    temporarily to load only the tokenizer.
-    """
-    # mlx-community 'gemma4_unified' text checkpoints (e.g. gemma-4-12B-it-4bit)
-    # need a dedicated loader: their model_type has no mlx-lm module and their
-    # multimodal weights must be dropped to load the language tower. The loader
-    # materializes weights eagerly and ignores kwargs like ``lazy`` — these
-    # checkpoints are dense text towers, so the lazy flash-MoE caller never
-    # routes here.
-    gemma4_unified = _maybe_load_gemma4_unified_text(str(load_path))
-    if gemma4_unified is not None:
-        return gemma4_unified
-
-    _sanitize_model_config_in_place(load_path)
-    kwargs.setdefault("tokenizer_config", {"trust_remote_code": True})
-    try:
-        model, tokenizer = mlx_lm.load(str(load_path), **kwargs)
-    except (AttributeError, ValueError, KeyError) as exc:
-        config_file = Path(load_path) / "config.json"
-        if not config_file.exists():
-            raise
-        from transformers.models.auto.configuration_auto import CONFIG_MAPPING
-
-        original_text = config_file.read_text()
-        cfg = json.loads(original_text)
-        mt = cfg.get("model_type", "")
-        # Strip last digit: deepseek_v32 → deepseek_v3
-        fallback = re.sub(r"\d+$", lambda m: m.group()[:-1], mt) if mt else ""
-        if not fallback or fallback == mt or fallback not in CONFIG_MAPPING:
-            raise
-
-        logger.info(
-            "Loading model with real config, tokenizer with %r -> %r (%s)",
-            mt,
-            fallback,
-            exc,
-        )
-        # Load model with the real config (mlx-lm knows the architecture).
-        model, model_cfg = mlx_lm.utils.load_model(
-            Path(load_path),
-            **{k: v for k, v in kwargs.items() if k in ("lazy", "model_config")},
-        )
-        # Temporarily patch config.json for tokenizer loading only.
-        try:
-            cfg["model_type"] = fallback
-            config_file.write_text(json.dumps(cfg, indent=2, ensure_ascii=False))
-            tokenizer = mlx_lm.utils.load_tokenizer(
-                Path(load_path),
-                kwargs.get("tokenizer_config"),
-                eos_token_ids=model_cfg.get("eos_token_id", None),
-            )
-        finally:
-            config_file.write_text(original_text)
-    # Outside try/except: an AttributeError from the EOS helper must not
-    # trigger the model-type remapping fallback with a misleading error.
-    _ensure_tokenizer_eos_in_stops(tokenizer)
-    # Materialize non-parameter buffers (scaled-RoPE ``_freqs``, ...) on THIS
-    # (load) thread — mlx-lm's ``mx.eval(model.parameters())`` skips them, and
-    # left lazy they crash when first evaluated on the generation worker thread
-    # (#499 thread-local streams). Covers both the mlx-lm happy path and the
-    # config-remap fallback above.
-    _materialize_module_buffers(model)
-    return model, tokenizer
-
-
-def _quantize_language_tower(
-    model: Any, quant: dict, lang_weights: dict, prefix: str = "language_model."
-) -> None:
-    """Quantize a rebuilt ``gemma4_text`` language tower in place, honoring the
-    checkpoint's *per-layer* quantization overrides.
-
-    mlx-community ``gemma4_unified`` checkpoints carry mixed precision: a global
-    ``bits``/``group_size`` plus per-module overrides (e.g. MLP projections at
-    8-bit) keyed by the full ``language_model.*`` path.  A blanket
-    ``nn.quantize`` at the global bits builds QuantizedLinear params whose packed
-    shapes mismatch the stored 8-bit weights, so the strict load fails.  This
-    mirrors mlx-lm's own ``load_model`` predicate, mapping our stripped tower
-    paths (``model.layers.0.mlp.gate_proj``) back to the prefixed config keys.
-    """
-    import mlx.nn as nn
-
-    def class_predicate(path: str, module: Any) -> bool | dict:
-        # Per-layer override wins: config keys retain the language_model prefix.
-        override = quant.get(f"{prefix}{path}")
-        if isinstance(override, dict):
-            return override
-        if not hasattr(module, "to_quantized"):
-            return False
-        # Otherwise quantize at the global bits only if the checkpoint stored
-        # this module as quantized (a non-quantized module has no .scales).
-        return f"{path}.scales" in lang_weights
-
-    nn.quantize(
-        model,
-        group_size=quant.get("group_size", 64),
-        bits=quant.get("bits", 4),
-        mode=quant.get("mode", "affine"),
-        class_predicate=class_predicate,
-    )
-
-
-def _load_gemma4_unified_text(load_path: str, config: dict) -> tuple[Any, Any]:
-    """Load only the *language tower* of a mlx-community ``gemma4_unified``
-    checkpoint (e.g. ``gemma-4-12B-it-4bit``) via mlx-lm's ``gemma4_text``.
-
-    These repos ship the full multimodal checkpoint under a non-standard
-    ``gemma4_unified`` model_type whose vision tower is stored under a
-    ``vision_embedder.*`` prefix.  That layout matches neither mlx-lm's
-    ``gemma4_text`` module (it wants the language tower under ``model.*``) nor
-    mlx-vlm 0.4.4's ``gemma4`` module (it models the vision tower as
-    ``vision_tower`` + ``embed_vision``, so the 11 ``vision_embedder.*`` params
-    have no home).  For text inference we build ``gemma4_text`` from
-    ``text_config``, load the ``language_model.*`` weights (stripping the
-    prefix), and drop the multimodal weights.  Standard ``gemma4`` checkpoints
-    (e4b/31b) are unaffected and keep their normal mlx-vlm routing.
-    """
-    import glob
-
-    import mlx.core as mx
-    import mlx_lm
-    from mlx_lm.models import gemma4_text
-
-    text_cfg = dict(config.get("text_config", {}))
-    # gemma4_text's own module_type; the checkpoint carries
-    # "gemma4_unified_text" which the mlx-lm ModelArgs doesn't recognise.
-    text_cfg["model_type"] = "gemma4_text"
-    args = gemma4_text.ModelArgs.from_dict(text_cfg)
-    model = gemma4_text.Model(args)
-
-    weights: dict[str, Any] = {}
-    for shard in sorted(glob.glob(str(Path(load_path) / "*.safetensors"))):
-        # mx.load is typed Union[array, dict, tuple]; the dict case is what
-        # safetensors returns. Same suppression as pre_shard.py / flash loaders.
-        weights.update(mx.load(shard))  # pyright: ignore[reportCallIssue]
-    # Keep only the language tower; the multimodal weights (vision_embedder,
-    # embed_vision, embed_audio) are intentionally dropped for the text path.
-    prefix = "language_model."
-    lang_weights = {
-        k[len(prefix) :]: v for k, v in weights.items() if k.startswith(prefix)
-    }
-    if not lang_weights:
-        raise ValueError(
-            f"gemma4_unified checkpoint at {load_path} has no "
-            f"'{prefix}*' weights; cannot load the language tower."
-        )
-    lang_weights = model.sanitize(lang_weights)
-
-    # Quantize *before* loading so the packed param shapes match the stored
-    # weights.  These checkpoints carry mixed precision (a global bits plus
-    # per-module 8-bit overrides), so quantize from the per-layer config rather
-    # than a single blanket bits — sanitize first to mirror mlx-lm's order.
-    quant = config.get("quantization") or config.get("quantization_config")
-    if quant:
-        _quantize_language_tower(model, quant, lang_weights, prefix)
-
-    model.load_weights(list(lang_weights.items()), strict=True)
-    mx.eval(model.parameters())
-
-    # Thread the full eos_token_id list into the tokenizer's stop set.  Gemma 4
-    # terminates turns with ``<turn|>`` (id 106) and pauses for tool results at
-    # ``<|tool_response>`` (id 50) — both are stop tokens alongside ``<eos>``
-    # (id 1).  Without them generation runs past every turn/tool boundary and
-    # degenerates into a repetition loop.  generation_config.json wins over
-    # config.json (mirrors mlx-lm's load_config precedence).
-    eos_token_ids = config.get("eos_token_id")
-    gen_config_file = Path(load_path) / "generation_config.json"
-    if gen_config_file.exists():
-        try:
-            gen_config = json.loads(gen_config_file.read_text())
-        except (OSError, ValueError):
-            gen_config = {}
-        if gen_config.get("eos_token_id") is not None:
-            eos_token_ids = gen_config["eos_token_id"]
-
-    tokenizer = mlx_lm.utils.load_tokenizer(
-        Path(load_path),
-        {"trust_remote_code": True},
-        eos_token_ids=eos_token_ids,
-    )
-    _ensure_tokenizer_eos_in_stops(tokenizer)
-    return model, tokenizer
-
-
-def _maybe_load_gemma4_unified_text(load_path: str) -> tuple[Any, Any] | None:
-    """Return ``(model, tokenizer)`` for a text-routable ``gemma4_unified``
-    checkpoint, or ``None`` if this isn't one (so callers fall through to the
-    normal mlx-lm load).  Reads config.json only — never mutates it.
-    """
-    config_file = Path(load_path) / "config.json"
-    if not config_file.exists():
-        return None
-    try:
-        config = json.loads(config_file.read_text())
-    except (OSError, ValueError):
-        return None
-    if config.get("model_type", "").lower() != "gemma4_unified":
-        return None
-    text_model_type = config.get("text_config", {}).get("model_type", "").lower()
-    if text_model_type != "gemma4_unified_text":
-        return None
-    return _load_gemma4_unified_text(load_path, config)
-
-
-def _is_serializable_cache(cache: list) -> bool:
-    """Check if a cache list can be serialized with mlx-lm's save_prompt_cache."""
-    from olmlx.engine.shardquant_cache import ShardKVCache
-    from olmlx.engine.spectralquant_cache import SpectralQuantKVCache
-    from olmlx.engine.turboquant_cache import TurboQuantKVCache
-
-    return not any(
-        isinstance(c, (TurboQuantKVCache, SpectralQuantKVCache, ShardKVCache))
-        for c in cache
-    )
-
-
-# KV-quant config string prefixes whose backing cache class cannot be
-# deep-copied by ``snapshot_cache_for_persistence``.  Currently empty:
-# - ``turboquant:`` is handled by ``TurboQuantKVCache.__deepcopy__`` which
-#   shares the ``mx.Dtype`` singleton by reference and eager-evals the
-#   private dequant side buffers that ``flatten_cache_state`` doesn't see.
-# - ``spectral:`` deepcopies cleanly via the default walk (no ``mx.Dtype``
-#   attribute, no out-of-state side buffers).
-# - ``shard:`` likewise deepcopies via the default walk: ShardKVCache keeps
-#   no ``mx.Dtype`` attribute and exposes every mutable array (sink, window,
-#   packed middle, norms) through ``state``, so the flatten + ``mx.eval``
-#   pass materializes its in-place-write graphs before the copy.
-# Kept as the registration point for any future kv-quant cache class
-# that proves unsafe — mirrors ``_EXCLUDED_MIXED_LAYER_PAIRS``.  Disk-save
-# remains gated separately by ``_is_serializable_cache`` because that path
-# uses safetensors and the packed indices/codebook layout has no
-# upstream-compatible serialisation.
-_KV_QUANT_PREFIXES_BLOCKING_SNAPSHOT: tuple[str, ...] = ()
-
-
-def _kv_quant_blocks_snapshot(kv_cache_quant: str | None) -> bool:
-    """True iff ``kv_cache_quant`` selects a quantization wrapper whose
-    layer state can't be deep-copied (the checkpoint path requirement).
-
-    Prefix-based because the probe constructs the bare (unquantized) cache
-    by design — see the comment on ``_probe_cache_capabilities`` — so we
-    can't check via ``isinstance`` like ``_is_serializable_cache`` does.
-    """
-    if kv_cache_quant is None:
-        return False
-    return any(
-        kv_cache_quant.startswith(prefix)
-        for prefix in _KV_QUANT_PREFIXES_BLOCKING_SNAPSHOT
-    )
-
-
-# mlx-lm cache classes whose `trim(n)` reliably removes exactly n tokens from
-# any offset (no silent under-delivery).  Class-name allowlist is authoritative
-# because `is_trimmable()` on a fresh (offset==0) cache cannot detect rotate
-# or chunk problems that only manifest once the buffer fills.
-#
-# Verified trim behaviour against mlx-lm's cache.py:
-# - KVCache.trim(n) → min(offset, n); trims back to 0 cleanly (line 378).
-# - QuantizedKVCache.trim(n) → same clamping (line 309).
-# - ConcatenateKVCache.trim(n) → same clamping (line 214); used by afm7.
-#
-# Deliberately excluded:
-# - RotatingKVCache: is_trimmable() returns False once the ring buffer fills.
-# - ChunkedKVCache: trim() silently clamps to (offset - start_position);
-#   once prompt > chunk_size, trims beyond the current chunk under-deliver.
-# - ArraysCache / CacheList: no usable trim semantics for our purposes.
-# - BatchKVCache / BatchRotatingKVCache: batch path; untested here.
-#
-# Probe always inspects bare mlx-lm classes (not quantization wrappers like
-# TurboQuant/Spectral), so those wrappers do not need to appear here even
-# though they implement trim correctly — they're simply never the layer
-# types we see.  See ml-explore/mlx-lm#980 for upstream hybrid-trim work.
-_TRIMMABLE_CACHE_CLASSES = frozenset(
-    {
-        "KVCache",
-        "QuantizedKVCache",
-        "ConcatenateKVCache",
-    }
-)
-
-
-def _cache_supports_trim(cache_list: list) -> bool:
-    """True iff every layer is in the known-trimmable allowlist.
-
-    Pure check against ``_TRIMMABLE_CACHE_CLASSES``; no knowledge of the
-    persistence flag.  Used at model load time to decide whether
-    ``trim_prompt_cache()`` is worth attempting on this model.  Hybrid
-    sliding-window models (Gemma 4, Qwen3-Next, etc.) include
-    ``RotatingKVCache`` layers and return False.  The downstream
-    consequence — non-trimmable layouts also lose cross-request
-    persistence (#343) — is applied by ``_probe_cache_capabilities``,
-    not here.
-    """
-    return all(type(layer).__name__ in _TRIMMABLE_CACHE_CLASSES for layer in cache_list)
-
-
-# Cache layer classes whose stored state CAN be safely reused on a later
-# request.  Allowlist (not denylist) so that any new mlx-lm cache class
-# defaults to non-persistable: a false negative here just means a missed
-# cache hit, while a false positive (treating an unknown SSM-style cache
-# as persistable) crashes mlx-lm during the next prefill with "RuntimeError:
-# There is no Stream(gpu, N) in current thread".  Issue #284.
-#
-# ArraysCache (gated-delta SSM state used by Qwen3.5, Qwen3-Next) is the
-# motivating exclusion: its stored arrays carry a lazy graph that
-# references a Metal stream from the previous worker thread; re-evaluating
-# them in a fresh worker thread fails.  Within-request cache reuse during
-# a single generation works fine; only cross-request persistence is unsafe.
-#
-# Verified safe against mlx-lm by inspection of mlx_lm/models/cache.py:
-# the failure mode in #284 is that ArraysCache stores arrays produced by
-# the ``gated_delta_kernel`` (mx.fast.metal_kernel) whose lazy graph
-# carries a Metal stream reference from the generating worker thread,
-# and re-evaluating that graph in a different worker thread raises
-# "There is no Stream(gpu, N)".  The classes in the allowlist below
-# all store keys/values produced by stock matmul/attention ops only —
-# no metal_kernel outputs, no generator-thread-bound state — so their
-# arrays are reusable across worker threads:
-# - KVCache: bare keys + values, mx.concatenate at update boundaries.
-# - QuantizedKVCache: same, with mx.quantize/dequantize wrappers.
-# - ConcatenateKVCache: same, used by AFM-7 family.
-# - RotatingKVCache: ring buffer over fixed window (Gemma 4); writes
-#   in place via mx.assign at modular offsets.  Non-trimmable, so
-#   ``_probe_cache_capabilities`` folds this to effective persist=False
-#   (issue #343) — listed here only because the bare layout itself is
-#   Metal-safe to store.
-# - ChunkedKVCache: chunked layout (afm7); same semantics as KVCache
-#   bounded by chunk_size.  Also non-trimmable, same #343 fold applies.
-# If a future mlx-lm release adds metal_kernel-style state to any of
-# these, that class must be removed from the allowlist.
-#
-# Deliberately excluded:
-# - ArraysCache: gated-delta SSM state — see issue #284.
-# - CacheList: wraps other caches, would need recursion to classify safely.
-# - BatchKVCache / BatchRotatingKVCache: batch path, untested in olmlx
-#   (single-user server) — disable persistence rather than risk a crash.
-_PERSISTABLE_CACHE_CLASSES = frozenset(
-    {
-        "KVCache",
-        "QuantizedKVCache",
-        "ConcatenateKVCache",
-        "RotatingKVCache",
-        "ChunkedKVCache",
-    }
-)
-
-
-# Layer classes that are safe to persist via the message-boundary
-# CHECKPOINT path (not the flat path).  Strict superset of
-# _PERSISTABLE_CACHE_CLASSES: RotatingKVCache is safe because the
-# checkpoint mechanism avoids trim entirely (#343); ArraysCache is
-# safe when the snapshot helper materialises the lazy gated_delta_kernel
-# graph via mx.eval before storage (#284).
-_PERSISTABLE_CACHE_CLASSES_WITH_CHECKPOINT = frozenset(
-    {
-        "KVCache",
-        "QuantizedKVCache",
-        "ConcatenateKVCache",
-        "RotatingKVCache",
-        "ChunkedKVCache",
-        "ArraysCache",
-    }
-)
-
-# Layer-class compositions that are individually safe via the checkpoint
-# path but combine into a layout we can't currently snapshot consistently.
-# Empty since #396 was lifted: the Qwen3-Next mix (RotatingKVCache +
-# ArraysCache) is handled by ``snapshot_cache_for_persistence`` — each
-# layer is deepcopied independently and ArraysCache lazy state is
-# materialized via ``mx.eval`` before the deepcopy, so the joint snapshot
-# captures both sub-states at the same logical token offset and survives
-# a continued forward pass after restore.  Kept as the registration point
-# for any future composition that proves unsafe.
-_EXCLUDED_MIXED_LAYER_PAIRS: frozenset[frozenset[str]] = frozenset()
-
-
-def _layer_layout_is_mixed_excluded(cache_list: list) -> bool:
-    """True iff the layer-class set is in the excluded mixed-pair list."""
-    if not cache_list:
-        return False
-    classes = {type(layer).__name__ for layer in cache_list}
-    for excluded in _EXCLUDED_MIXED_LAYER_PAIRS:
-        if excluded.issubset(classes):
-            return True
-    return False
-
-
-def _cache_supports_checkpoint_persistence(cache_list: list) -> bool:
-    """True iff every layer is checkpoint-persistable AND the layout is
-    not in the excluded mixed-pair list.
-    """
-    if not cache_list:
-        return False
-    if _layer_layout_is_mixed_excluded(cache_list):
-        return False
-    return all(
-        type(layer).__name__ in _PERSISTABLE_CACHE_CLASSES_WITH_CHECKPOINT
-        for layer in cache_list
-    )
-
-
-def _cache_supports_persistence(cache_list: list) -> bool:
-    """True iff every layer is a cache type known to be safe for
-    cross-request persistence.
-
-    See ``_PERSISTABLE_CACHE_CLASSES`` for the allowlist.  Models that fail
-    this check still get within-request cache reuse — just not the
-    cross-request strict-extension reuse the prompt cache store normally
-    provides.
-
-    Exact class-name match (no MRO walk).  With allowlist semantics an MRO
-    walk would invert the safety guarantee: a future ``BadSSMCache(KVCache)``
-    that has unsafe state but inherits from an allowlisted class would
-    silently pass.  Subclasses must be added to the allowlist explicitly.
-    A false-negative just costs a cache miss; a false-positive crashes
-    mlx-lm with a Metal stream error on the next request.
-
-    Empty ``cache_list`` returns False (no evidence of safety).  Unlike the
-    trim probe — where a false-positive falls back gracefully — a stray
-    True here would crash the next request.
-    """
-    if not cache_list:
-        return False
-    for layer in cache_list:
-        if type(layer).__name__ not in _PERSISTABLE_CACHE_CLASSES:
-            return False
-    return True
-
-
-class ModelLoadTimeoutError(TimeoutError):
-    """Raised when model loading exceeds OLMLX_MODEL_LOAD_TIMEOUT."""
-
-
-class SpectralCalibrationMissingError(Exception):
-    """Raised when SpectralQuant is configured but calibration data is absent."""
-
-
-class ShardCalibrationMissingError(SpectralCalibrationMissingError):
-    """Shard quant configured but calibration artifacts missing/mismatched.
-
-    Subclasses SpectralCalibrationMissingError so the existing app.py
-    exception handler (HTTP 400) catches it via Starlette's MRO walk.
-    """
-
-
-class ActiveRequestsError(RuntimeError):
-    """Raised by ``ModelManager.unload`` when a model has in-flight requests.
-
-    Subclasses ``RuntimeError`` so legacy ``except RuntimeError:`` keeps
-    working, but the dedicated type lets HTTP routers narrow the 409 path
-    to exactly this condition. Without it, an unrelated ``RuntimeError``
-    from ``_close_loaded_model`` would be misreported as 409.
-    """
-
-
-def structural_copy(module: "nn.Module") -> "nn.Module":
-    """Copy an ``nn.Module`` *tree* while sharing every weight by reference.
-
-    LoRA hot-swap (issue #362) needs one base model's weights resident once in
-    RAM with multiple adapters layered on top. ``mlx_lm.tuner.utils.load_adapters``
-    mutates the module tree in place (replacing ``nn.Linear`` with ``LoRALinear``),
-    so each adapter needs its *own* module objects — but the heavy ``mx.array``
-    weights must stay shared.
-
-    Neither ``copy.copy`` nor ``copy.deepcopy`` works:
-
-    * ``copy.copy`` is shallow — the ``layers`` list and every submodule are
-      shared by reference, so ``load_adapters`` would corrupt the base model.
-    * ``copy.deepcopy`` duplicates the ``mx.array`` weights (defeating sharing)
-      and crashes on quantized models — modules store an ``mx.Dtype`` attribute
-      and ``Dtype`` is not picklable.
-
-    So this walks the tree explicitly: fresh ``Module`` / list / dict / tuple
-    containers, with ``mx.array`` leaves and scalar attributes (``bits``,
-    ``group_size``, dtypes, …) shared by reference. Mutable per-module Python
-    state (the ``_no_grad`` / freeze-tracking sets) is copied so freezing one
-    instance can't affect another. ``nn.Module`` overrides ``__setattr__`` to
-    route into its backing dict, so attributes are copied by mutating
-    ``__dict__`` in place rather than reassigning it. Verified to share weights
-    (no extra Metal allocation) and isolate the base through a real
-    ``load_adapters`` call on a 4-bit-quantized model.
-
-    A by-id memo makes the walk robust to module graphs that aren't strict
-    trees: a submodule instance shared across layers (e.g. one rotary-embedding
-    object) is copied once and re-shared, and any back-reference / cycle
-    resolves to the in-progress copy instead of recursing forever. ``id()`` is
-    stable here because the whole source tree stays referenced for the walk's
-    duration, so no object is collected and no address is recycled.
-    """
-
-    memo: dict[int, Any] = {}
-
-    def _copy(obj: Any) -> Any:
-        oid = id(obj)
-        cached = memo.get(oid)
-        if cached is not None:
-            return cached
-        if isinstance(obj, nn.Module):
-            new = obj.__class__.__new__(obj.__class__)
-            memo[oid] = new  # register before recursing so cycles terminate
-            for k, v in obj.__dict__.items():
-                new.__dict__[k] = v.copy() if isinstance(v, (set, dict, list)) else v
-            for k, v in obj.items():
-                dict.__setitem__(new, k, _copy(v))
-            return new
-        if isinstance(obj, dict):
-            new_dict: dict = {}
-            memo[oid] = new_dict
-            for k, v in obj.items():
-                new_dict[k] = _copy(v)
-            return new_dict
-        if isinstance(obj, list):
-            new_list: list = []
-            memo[oid] = new_list
-            for v in obj:
-                new_list.append(_copy(v))
-            return new_list
-        if isinstance(obj, tuple):
-            # Immutable, so it can't be pre-registered before its contents are
-            # built — but any reference cycle reaching a tuple must pass through
-            # a mutable container (list/dict/Module), which IS pre-registered,
-            # so recursion still terminates. Register after building to dedupe a
-            # tuple shared from several places.
-            new_tuple = tuple(_copy(v) for v in obj)
-            memo[oid] = new_tuple
-            return new_tuple
-        return obj
-
-    return _copy(module)
-
-
-@dataclass
-class LoadedModel:
-    name: str
-    hf_path: str
-    model: Any
-    tokenizer: Any  # tokenizer (mlx-lm) or processor (mlx-vlm)
-    is_vlm: bool = False
-    is_distributed: bool = False
-    is_flash: bool = False
-    is_flash_moe: bool = False
-    is_whisper: bool = False
-    is_tts: bool = False
-    is_reranker: bool = False
-    speculative_decoder: Any = None
-    weight_store: Any = None
-    # LoRA-adapter hot-swap (issue #362). On an adapter entry, ``adapter_base``
-    # names the base model's key in ``_loaded`` (whose weights this entry
-    # shares via a structural copy); None on ordinary base models.
-    # ``_adapter_child_refs`` counts adapters currently sharing THIS entry's
-    # weights — a base with refs > 0 is pinned against LRU eviction. Guarded by
-    # the manager's ``_lock`` (mutated only on the loop thread).
-    adapter_base: str | None = None
-    _adapter_child_refs: int = field(default=0, compare=False, repr=False)
-    # Continuous-batching scheduler (engine/batching.py), created lazily on
-    # the first batch-eligible request. None for ineligible/idle models.
-    batch_scheduler: Any = None
-    # Memoized result of the batch cache-probe (None = not yet probed).
-    batch_convertible: bool | None = None
-    template_caps: TemplateCaps = field(default_factory=TemplateCaps)
-    loaded_at: float = field(default_factory=time.time)
-    expires_at: float | None = None
-    size_bytes: int = 0
-    active_refs: int = 0
-    _active_refs_lock: threading.Lock = field(
-        default_factory=threading.Lock, compare=False, repr=False
-    )
-    prompt_cache_store: PromptCacheStore = field(default=None)  # type: ignore[assignment]
-    # Per-cache_id LRU of mlx_vlm PromptCacheState objects for cross-turn
-    # image-prefix KV reuse. Only populated for VLMs (None otherwise). #429.
-    vlm_prompt_cache_store: Any = None
-    kv_cache_quant: str | None = None
-    #: StreamingLLM sink+window KV eviction "<sink>:<window>" (#505), or None.
-    kv_eviction: str | None = None
-    #: Weight quantization string (e.g. "hqq:4") applied at load time.
-    #: ``None`` means no weight quantization was applied.
-    weight_quant: str | None = None
-    # False for hybrid sliding-window models (RotatingKVCache layers).
-    # Set by the loader; default True covers direct construction in tests.
-    supports_cache_trim: bool = True
-    # False for hybrid SSM-style models (ArraysCache layers, e.g. Qwen3.5,
-    # Qwen3-Next).  When False, prompt cache state is not persisted across
-    # requests because cross-request reuse crashes mlx-lm with a Metal
-    # stream error.  Set by the loader's _probe_cache_capabilities call.
-    # Issue #284.
-    #
-    # Defaults to False (unsafe-by-default) — unlike supports_cache_trim,
-    # a false-positive here crashes the next request rather than wasting a
-    # trim_prompt_cache() call.  Direct LoadedModel construction must
-    # explicitly opt in if the model's cache layout is known to be safe.
-    supports_cache_persistence: bool = False
-    # True for models whose cache layout is non-trimmable (ArraysCache,
-    # RotatingKVCache) but can still benefit from cross-request reuse via
-    # the checkpoint path: prefill per message-segment, snapshot at each
-    # boundary.  Set by _probe_cache_capabilities (Tasks 5.2/5.3).
-    # Defaults to False until the probe populates it.
-    uses_checkpoint_persistence: bool = False
-    spectral_calibration_dir: Any = None  # Path | None, typed as Any to avoid import
-    shard_calibration_dir: Any = None  # Path | None, typed as Any to avoid import
-    default_options: dict = field(default_factory=dict)
-    inference_queue_timeout: float | None = None
-    inference_timeout: float | None = None
-    sync_mode: SyncMode | None = None
-    # Per-model default for the chat-template ``enable_thinking`` kwarg.
-    # ``generate_chat`` consults this when the request didn't set the
-    # flag.  None means defer to the engine default.
-    enable_thinking: bool | None = None
-    # Per-model default reasoning level for channel-format reasoners
-    # (gpt-oss / Harmony): "low"/"medium"/"high". ``generate_chat`` consults
-    # this when the caller didn't pass ``reasoning_effort``.
-    reasoning_effort: str | None = None
-    # Per-model override for the cross-request prompt cache toggle.
-    # ``generate_chat`` honours this in place of ``settings.prompt_cache``
-    # when set. None means defer to the global setting. Surfaced for
-    # architectures that crash or degrade on the checkpoint path while
-    # other models continue to benefit from caching (e.g. Qwen3-Coder-Next
-    # MoE-quantized targets where chunked prefill triggers GatedDeltaNet
-    # numerical drift across expert-routing thresholds).
-    prompt_cache: bool | None = None
-    # Per-model override for continuous batching. ``_batch_eligible``
-    # consults this in place of ``settings.batching`` when set; None defers
-    # to the global toggle. Mechanical eligibility (cache layout, model
-    # kind, KV-quant, …) still applies on top. The batch-size knobs below
-    # likewise override ``settings.batch_*`` in ``_get_batch_scheduler``
-    # when set; None means inherit the global value.
-    batching: bool | None = None
-    batch_completion_size: int | None = None
-    batch_prefill_size: int | None = None
-    batch_prefill_step: int | None = None
-    batch_fairness_quantum: float | None = None
-
-    def __post_init__(self):
-        if self.prompt_cache_store is None:
-            disk_path = (
-                Path(settings.prompt_cache_disk_path).expanduser()
-                if settings.prompt_cache_disk
-                else None
-            )
-            disk_max_bytes = (
-                int(settings.prompt_cache_disk_max_gb * 1024**3)
-                if settings.prompt_cache_disk
-                else None
-            )
-            ram_budget_bytes = int(settings.prompt_cache_ram_budget_gb * 1024**3)
-            self.prompt_cache_store = PromptCacheStore(
-                max_slots=settings.prompt_cache_max_slots,
-                disk_path=disk_path,
-                model_name=self.name,
-                disk_max_bytes=disk_max_bytes,
-                ram_budget_bytes=ram_budget_bytes,
-            )
-        if self.is_vlm and self.vlm_prompt_cache_store is None:
-            from olmlx.engine.prompt_cache.vlm_state import VlmPromptCacheStore
-
-            vlm_disk_path = (
-                settings.vlm_prompt_cache_disk_path
-                if settings.vlm_prompt_cache_disk
-                else None
-            )
-            vlm_disk_max_bytes = (
-                int(settings.vlm_prompt_cache_disk_max_gb * 1024**3)
-                if settings.vlm_prompt_cache_disk
-                else None
-            )
-            self.vlm_prompt_cache_store = VlmPromptCacheStore(
-                capacity=settings.vlm_prompt_cache_slots,
-                disk_path=vlm_disk_path,
-                model_name=self.name,
-                disk_max_bytes=vlm_disk_max_bytes,
-            )
-
-    def acquire_ref(self) -> None:
-        """Pin the model so eviction/expiry skip it (active_refs += 1)."""
-        with self._active_refs_lock:
-            self.active_refs += 1
-
-    def release_ref(self) -> None:
-        """Release a pin taken by :meth:`acquire_ref` (active_refs -= 1)."""
-        with self._active_refs_lock:
-            self.active_refs -= 1
-
-    @property
-    def is_speculative(self) -> bool:
-        return self.speculative_decoder is not None
-
-    @property
-    def text_tokenizer(self) -> Any:
-        """Return the underlying text tokenizer, unwrapping VLM processor if needed.
-
-        mlx-vlm's load() returns a processor whose .tokenizer attribute is
-        the actual HuggingFace tokenizer with the chat template.
-        """
-        tok = self.tokenizer
-        if self.is_vlm and hasattr(tok, "tokenizer"):
-            return tok.tokenizer
-        return tok
-
-
-def _is_cross_encoder_config(config: dict) -> bool:
-    """True for an XLM-RoBERTa-family sequence-classification reranker (#369).
-
-    Requires both a ``*ForSequenceClassification`` architecture AND a
-    roberta-family ``model_type`` — the encoder and weight remaps only support
-    XLM-RoBERTa/RoBERTa, so a plain BERT/DistilBERT text classifier must NOT be
-    routed here (it would otherwise be detected as a reranker and fail at load).
-    """
-    archs = config.get("architectures") or []
-    if not any(
-        isinstance(a, str) and a.endswith("ForSequenceClassification") for a in archs
-    ):
-        return False
-    return "roberta" in (config.get("model_type") or "").lower()
-
-
-def parse_keep_alive(value: str | int) -> float | None:
-    """Parse keep_alive to seconds. Returns None for never-expire (-1)."""
-    if isinstance(value, (int, float)):
-        if value < 0:
-            return None
-        return float(value)
-    value = str(value).strip()
-    if value == "-1":
-        return None
-    if value == "0":
-        return 0.0
-    # Bare integer string → treat as seconds (consistent with Ollama API)
-    if value.isdigit():
-        return float(int(value))
-    match = re.match(r"^(\d+)(s|m|h)$", value)
-    if not match:
-        logger.warning("Invalid keep_alive format: %r, defaulting to 5m", value)
-        return 300.0  # default 5m
-    num, unit = int(match.group(1)), match.group(2)
-    multipliers = {"s": 1, "m": 60, "h": 3600}
-    return float(num * multipliers[unit])
 
 
 class ModelManager(SpeculativeLoaderMixin):
